+++ src/deployment.ts
@@ -154,7 +154,7 @@ export class Deployment extends LongRunningWorkload {
       replicas: this.replicas,
       template: {
         metadata: this.podMetadata.toJson(),
-        spec: this.podSpec._toKube(),
+        spec: this.podSpec,
       },
       selector: {
         matchLabels: this.labelSelector,
